Fix high cpu consumption
when running for a long time, there are 5000 threads in the collection.
Each time we update the collection, it incurs a high cpu consumption.
The update of the collection is broken: we remove the "min". But the
newly added object, is the new "min". Which means that the collection,
doesn't have the new objects.